requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.4
django-autoslug==1.9.8
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-widget-tweaks==1.4.8
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
Pillow==8.0.1
psycopg2==2.8.6
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2020.4
reportlab==3.5.56
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
webencodings==0.5.1
whitenoise==5.1.0
xhtml2pdf==0.2.5
razorpay==1.2.0
twilio==6.51.1